requirements.txt
flask==2.3.3
flask-restful==0.3.10
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.3
pyjwt==2.8.0
openai==0.28.0
rich==13.5.3
prompt-toolkit==3.0.39
pymetasploit3==1.0.0
psutil==5.9.5
pyyaml==6.0.1
